@@ -1,4 +1,6 @@
-class AddCurrentSignInAtToUsers < ActiveRecord::Migration
+# frozen_string_literal: true
+
+class AddCurrentSignInAtToUsers < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :current_sign_in_at, :datetime
   end
