@@ -1,8 +1,10 @@
+# frozen_string_literal: true
+
 # This file is copied to spec/ when you run 'rails generate rspec:install'
 ENV['RAILS_ENV'] ||= 'test'
 require 'dummy/config/environment'
 # Prevent database truncation if the environment is production
-abort("The Rails environment is running in production mode!") if Rails.env.production?
+abort('The Rails environment is running in production mode!') if Rails.env.production?
 require 'spec_helper'
 require 'rspec/rails'
 # Add additional requires below this line. Rails is not loaded until this point!
@@ -24,15 +26,14 @@ require 'rspec/rails'
 
 # Checks for pending migration and applies them before tests are run.
 # If you are not using ActiveRecord, you can remove this line.
-ActiveRecord::Migration.maintain_test_schema!
-
-# Remove after dropping support of Rails 4.2
-require_relative 'support/http_method_shim.rb'
+Dir.chdir('spec/dummy') do
+  ActiveRecord::Migration.maintain_test_schema!
+end
 
 require_relative 'support/doorkeeper_configuration.rb'
 
-require 'factory_girl'
-FactoryGirl.find_definitions
+require 'factory_bot'
+FactoryBot.find_definitions
 
 RSpec.configure do |config|
   # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
@@ -48,11 +49,11 @@ RSpec.configure do |config|
   # arbitrary gems may also be filtered via:
   # config.filter_gems_from_backtrace("gem name")
 
-  config.include FactoryGirl::Syntax::Methods
+  config.include FactoryBot::Syntax::Methods
 
   # Reinitialize configuration after each example
-  config.after :each do
-    load "#{Rails.root}/config/initializers/doorkeeper.rb"
-    load "#{Rails.root}/config/initializers/doorkeeper_openid_connect.rb"
+  config.after do
+    load Rails.root.join('config/initializers/doorkeeper.rb')
+    load Rails.root.join('config/initializers/doorkeeper_openid_connect.rb')
   end
 end
