@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Doorkeeper
   module OpenidConnect
     module AccessGrant
@@ -5,12 +7,11 @@ module Doorkeeper
         base.class_eval do
           has_one :openid_request,
             class_name: 'Doorkeeper::OpenidConnect::Request',
+            foreign_key: 'access_grant_id',
             inverse_of: :access_grant,
             dependent: :delete
         end
       end
     end
   end
-
-  AccessGrant.send :prepend, OpenidConnect::AccessGrant
 end
