@@ -1,5 +1,7 @@
+# frozen_string_literal: true
+
 module Doorkeeper
   module OpenidConnect
-    VERSION = '1.5.3'.freeze
+    VERSION = '1.8.7'
   end
 end
