@@ -1,20 +1,23 @@
+# frozen_string_literal: true
+
 require 'rails_helper'
 
 describe Doorkeeper::OpenidConnect, 'configuration' do
-  subject { Doorkeeper::OpenidConnect.configuration }
+  subject { described_class.configuration }
 
   describe '#configure' do
     it 'fails if not set to :active_record' do
       # stub ORM setup to avoid Doorkeeper exceptions
       allow(Doorkeeper).to receive(:setup_orm_adapter)
       allow(Doorkeeper).to receive(:setup_orm_models)
+      allow(Doorkeeper).to receive(:setup_application_owner)
 
       Doorkeeper.configure do
         orm :mongoid
       end
 
       expect do
-        Doorkeeper::OpenidConnect.configure {}
+        described_class.configure {}
       end.to raise_error Doorkeeper::OpenidConnect::Errors::InvalidConfiguration
     end
   end
@@ -22,7 +25,7 @@ describe Doorkeeper::OpenidConnect, 'configuration' do
   describe 'jws_private_key' do
     it 'delegates to signing_key' do
       value = 'private_key'
-      Doorkeeper::OpenidConnect.configure do
+      described_class.configure do
         jws_private_key value
       end
       expect(subject.signing_key).to eq(value)
@@ -32,7 +35,7 @@ describe Doorkeeper::OpenidConnect, 'configuration' do
   describe 'signing_key' do
     it 'sets the value that is accessible via signing_key' do
       value = 'private_key'
-      Doorkeeper::OpenidConnect.configure do
+      described_class.configure do
         signing_key value
       end
       expect(subject.signing_key).to eq(value)
@@ -42,24 +45,32 @@ describe Doorkeeper::OpenidConnect, 'configuration' do
   describe 'issuer' do
     it 'sets the value that is accessible via issuer' do
       value = 'issuer'
-      Doorkeeper::OpenidConnect.configure do
+      described_class.configure do
         issuer value
       end
       expect(subject.issuer).to eq(value)
+    end
+
+    it 'sets the block that is accessible via issuer' do
+      block = proc {}
+      described_class.configure do
+        issuer(&block)
+      end
+      expect(subject.issuer).to eq(block)
     end
   end
 
   describe 'resource_owner_from_access_token' do
     it 'sets the block that is accessible via resource_owner_from_access_token' do
       block = proc {}
-      Doorkeeper::OpenidConnect.configure do
+      described_class.configure do
         resource_owner_from_access_token(&block)
       end
       expect(subject.resource_owner_from_access_token).to eq(block)
     end
 
     it 'fails if unset' do
-      Doorkeeper::OpenidConnect.configure {}
+      described_class.configure {}
 
       expect do
         subject.resource_owner_from_access_token.call
@@ -70,14 +81,14 @@ describe Doorkeeper::OpenidConnect, 'configuration' do
   describe 'auth_time_from_resource_owner' do
     it 'sets the block that is accessible via auth_time_from_resource_owner' do
       block = proc {}
-      Doorkeeper::OpenidConnect.configure do
+      described_class.configure do
         auth_time_from_resource_owner(&block)
       end
       expect(subject.auth_time_from_resource_owner).to eq(block)
     end
 
     it 'fails if unset' do
-      Doorkeeper::OpenidConnect.configure {}
+      described_class.configure {}
 
       expect do
         subject.auth_time_from_resource_owner.call
@@ -88,14 +99,14 @@ describe Doorkeeper::OpenidConnect, 'configuration' do
   describe 'reauthenticate_resource_owner' do
     it 'sets the block that is accessible via reauthenticate_resource_owner' do
       block = proc {}
-      Doorkeeper::OpenidConnect.configure do
+      described_class.configure do
         reauthenticate_resource_owner(&block)
       end
       expect(subject.reauthenticate_resource_owner).to eq(block)
     end
 
     it 'fails if unset' do
-      Doorkeeper::OpenidConnect.configure {}
+      described_class.configure {}
 
       expect do
         subject.reauthenticate_resource_owner.call
@@ -103,17 +114,35 @@ describe Doorkeeper::OpenidConnect, 'configuration' do
     end
   end
 
+  describe 'select_account_for_resource_owner' do
+    it 'sets the block that is accessible via select_account_for_resource_owner' do
+      block = proc {}
+      described_class.configure do
+        select_account_for_resource_owner(&block)
+      end
+      expect(subject.select_account_for_resource_owner).to eq(block)
+    end
+
+    it 'fails if unset' do
+      described_class.configure {}
+
+      expect do
+        subject.select_account_for_resource_owner.call
+      end.to raise_error Doorkeeper::OpenidConnect::Errors::InvalidConfiguration
+    end
+  end
+
   describe 'subject' do
     it 'sets the block that is accessible via subject' do
       block = proc {}
-      Doorkeeper::OpenidConnect.configure do
+      described_class.configure do
         subject(&block)
       end
       expect(subject.subject).to eq(block)
     end
 
     it 'fails if unset' do
-      Doorkeeper::OpenidConnect.configure {}
+      described_class.configure {}
 
       expect do
         subject.subject.call
@@ -124,7 +153,7 @@ describe Doorkeeper::OpenidConnect, 'configuration' do
   describe 'expiration' do
     it 'sets the value that is accessible via expiration' do
       value = 'expiration'
-      Doorkeeper::OpenidConnect.configure do
+      described_class.configure do
         expiration value
       end
       expect(subject.expiration).to eq(value)
@@ -133,11 +162,11 @@ describe Doorkeeper::OpenidConnect, 'configuration' do
 
   describe 'claims' do
     it 'sets the claims configuration that is accessible via claims' do
-      Doorkeeper::OpenidConnect.configure do
+      described_class.configure do
         claims do
         end
       end
-      expect(subject.claims).to_not be_nil
+      expect(subject.claims).not_to be_nil
     end
   end
 
@@ -155,11 +184,49 @@ describe Doorkeeper::OpenidConnect, 'configuration' do
     end
 
     it 'can be set to other protocols' do
-      Doorkeeper::OpenidConnect.configure do
+      described_class.configure do
         protocol { :ftp }
       end
 
       expect(subject.protocol.call).to eq(:ftp)
+    end
+  end
+
+  describe 'end_session_endpoint' do
+    it 'defaults to nil' do
+      expect(subject.end_session_endpoint.call).to be_nil
+    end
+
+    it 'can be set to a custom url' do
+      described_class.configure do
+        end_session_endpoint { 'http://test.host/logout' }
+      end
+
+      expect(subject.end_session_endpoint.call).to eq('http://test.host/logout')
+    end
+  end
+
+  describe 'discovery_url_options' do
+    it 'defaults to empty hash' do
+      expect(subject.discovery_url_options.call).to be_kind_of(Hash)
+      expect(subject.discovery_url_options.call).to be_empty
+    end
+
+    it 'can be set to other hosts' do
+      Doorkeeper::OpenidConnect.configure do
+        discovery_url_options do |request|
+          {
+            authorization: { host: 'alternate-authorization-host' },
+            token: { host: 'alternate-token-host' },
+            revocation: { host: 'alternate-revocation-host' },
+            introspection: { host: 'alternate-introspection-host' },
+            userinfo: { host: 'alternate-userinfo-host' },
+            jwks: { host: 'alternate-jwks-host' }
+          }
+        end
+      end
+
+      expect(subject.discovery_url_options.call[:authorization]).to eq(host: 'alternate-authorization-host')
     end
   end
 end
