@@ -1,13 +1,16 @@
+# frozen_string_literal: true
+
 require 'rails_helper'
 
 describe Doorkeeper::OpenidConnect::IdToken do
   subject { described_class.new(access_token, nonce) }
+
   let(:access_token) { create :access_token, resource_owner_id: user.id, scopes: 'openid' }
   let(:user) { create :user }
   let(:nonce) { '123456' }
 
   before do
-    allow(Time).to receive(:now) { Time.at 60 }
+    allow(Time).to receive(:now) { Time.zone.at 60 }
   end
 
   describe '#nonce' do
@@ -18,7 +21,7 @@ describe Doorkeeper::OpenidConnect::IdToken do
 
   describe '#claims' do
     it 'returns all default claims' do
-      expect(subject.claims).to eq({
+      expect(subject.claims).to eq(
         iss: 'dummy',
         sub: user.id.to_s,
         aud: access_token.application.uid,
@@ -28,7 +31,27 @@ describe Doorkeeper::OpenidConnect::IdToken do
         auth_time: 23,
         both_responses: 'both',
         id_token_response: 'id_token',
-      })
+      )
+    end
+
+    context 'when application is not set on the access token' do
+      before do
+        access_token.application = nil
+      end
+
+      it 'returns all default claims except audience' do
+        expect(subject.claims).to eq(
+          iss: 'dummy',
+          sub: user.id.to_s,
+          aud: nil,
+          exp: 180,
+          iat: 60,
+          nonce: nonce,
+          auth_time: 23,
+          both_responses: 'both',
+          id_token_response: 'id_token',
+        )
+      end
     end
   end
 
@@ -40,8 +63,8 @@ describe Doorkeeper::OpenidConnect::IdToken do
 
       json = subject.as_json
 
-      expect(json).to_not include :iss
-      expect(json).to_not include :sub
+      expect(json).not_to include :iss
+      expect(json).not_to include :sub
       expect(json).to include :aud
     end
   end
@@ -49,8 +72,13 @@ describe Doorkeeper::OpenidConnect::IdToken do
   describe '#as_jws_token' do
     shared_examples 'a jws token' do
       it 'returns claims encoded as JWT' do
-        jwt = JSON::JWT.decode_compact_serialized subject.as_jws_token, Doorkeeper::OpenidConnect.signing_key
-        expect(jwt.to_hash).to eq subject.as_json.stringify_keys
+        algorithms = [Doorkeeper::OpenidConnect.signing_algorithm.to_s]
+
+        data, headers = ::JWT.decode subject.as_jws_token, Doorkeeper::OpenidConnect.signing_key.keypair, true, { algorithms: algorithms }
+
+        expect(data.to_hash).to eq subject.as_json.stringify_keys
+        expect(headers["kid"]).to eq Doorkeeper::OpenidConnect.signing_key.kid
+        expect(headers["alg"]).to eq Doorkeeper::OpenidConnect.signing_algorithm.to_s
       end
     end
 
