@@ -1,7 +1,10 @@
+# frozen_string_literal: true
+
 require 'rails_helper'
 
 describe Doorkeeper::OpenidConnect::OAuth::Authorization::Code do
   subject { Doorkeeper::OAuth::Authorization::Code.new pre_auth, resource_owner }
+
   let(:resource_owner) { create :user }
   let(:access_grant) { create :access_grant }
   let(:pre_auth) { double }
@@ -10,10 +13,13 @@ describe Doorkeeper::OpenidConnect::OAuth::Authorization::Code do
   describe '#issue_token' do
     before do
       allow(pre_auth).to receive(:client) { client }
-      allow(pre_auth).to receive(:redirect_uri) { 'redirect_uri' }
-      allow(pre_auth).to receive(:scopes) { 'scopes' }
-      allow(pre_auth).to receive(:nonce) { '123456' }
-      allow(client).to receive(:id) { 'client_id' }
+      allow(pre_auth).to receive(:redirect_uri).and_return('redirect_uri')
+      allow(pre_auth).to receive(:scopes).and_return('scopes')
+      allow(pre_auth).to receive(:nonce).and_return('123456')
+      allow(pre_auth).to receive(:code_challenge).and_return('987654')
+      allow(pre_auth).to receive(:code_challenge_method).and_return('plain')
+      allow(pre_auth).to receive(:custom_access_token_attributes).and_return({})
+      allow(client).to receive(:id).and_return('client_id')
 
       allow(Doorkeeper::AccessGrant).to receive(:create!) { access_grant }
       allow(Doorkeeper::OpenidConnect::Request).to receive(:create!)
@@ -29,17 +35,17 @@ describe Doorkeeper::OpenidConnect::OAuth::Authorization::Code do
     end
 
     it 'does not store the nonce if not present' do
-      allow(pre_auth).to receive(:nonce) { nil }
+      allow(pre_auth).to receive(:nonce).and_return(nil)
       subject.issue_token
 
-      expect(Doorkeeper::OpenidConnect::Request).to_not have_received(:create!)
+      expect(Doorkeeper::OpenidConnect::Request).not_to have_received(:create!)
     end
 
     it 'does not store the nonce if blank' do
-      allow(pre_auth).to receive(:nonce) { ' ' }
+      allow(pre_auth).to receive(:nonce).and_return(' ')
       subject.issue_token
 
-      expect(Doorkeeper::OpenidConnect::Request).to_not have_received(:create!)
+      expect(Doorkeeper::OpenidConnect::Request).not_to have_received(:create!)
     end
 
     it 'returns the created grant' do
